refactor: move deploy utils to lib
